refactor(core): move mute sync cursor clear helper to libs
